Add CertificateSubject optional parameter
This change will introduce an optional paramter to select a certificate
by filtering what is available in the local machine certificate store
and selecting a valid cert (by date).  If multiple certificates have the
same subject, the newest is selected.  If the parameter has a value and
a certificate is identified, it will overwrite the value of
CertificateThumbprint with the thumbprint of the selected certificate.
The documentation will will need to be updated to reflect this change.