Renamed namespace of MsiTools class
The original Package resource and xPackage were using the same namespace / class names for MsiTools, but the text isn't identical.  This was causing errors in Add-Type if a configuration happened to be using both Package and xPackage, when the second resource to run would try to add a class with the same name.