Adding xRegistry resource
This started with a copy of the Registry resource from PSDesiredStateConfiguration v1.1 (WMF 5.0.10586.0 from Windows 10).

Aside from the standard file renames for an "x" resource, this has tweaked the original Registry code slightly in order to support registry key names which contain a forward slash, such as 'HKLM:\SYSTEM\CurrentControlSet\Control\SecurityProviders\SCHANNEL\Ciphers\DES 56/56'.

The Registry PSProvider is a bit wonky for these paths, as it tries to treat the forward slash as a path separator instead of a name character.  New-Item will outright fail to create these keys, and Get-Item can also become confused if ambiguous paths exist. ('Ciphers\DES 56\56' and 'Ciphers\DES 56/56', for example.)  To remove this problem and avoid any other ambiguity, the code now uses the OpenSubKey and CreateSubKey methods on the Microsoft.Win32.RegistryKey class, instead of Get-Item and New-Item, respectively.