Added size comparison
When MatchSource is $False and Test-Path is $True added a webclient
content-length check to compare the size in bytes of the target and
destination files.

This will ensure that the file is re-downloaded if the file sizes are
not the same (i.e. target file has been updated, download was
interrupted previously, local file was corrupted, etc.)